First shot at a LaTeX output filter
This mostly works.  It does a good job of extracting errors and
warnings, but tends to over-consume input and miss the file name
stack.  Unfortunately, this is based on parsing the terminal output of
tex, rather than the log, and it turns out the terminal output has
many ambiguities that the log file does not.