Make BibTeX .aux file dependency more precise
Previously, we considered the .aux file a regular input to the bibtex
task, so any change would cause a re-run.  But bibtex actually depends
on very little of the .aux file, and other parts of it do tend to
change, so this was causing needless re-runs of bibtex.  Instead, pull
out and depend on only the lines that matter to BibTeX.