Pre-hash all output files to avoid extra latex runs
Previously, we only pre-hashed files that were both input and output
files, but this often resulted in three runs of latex when two would
suffice.  On the first run, latex wouldn't report the missing input
file, so we would think it was only an output.  On the second run,
latex would report that it also input the file, but we wouldn't know
that until after latex had run, at which point it was too late to
pre-hash the file.  Only on the third run would we know the file was
input/output and pre-hash it.

Instead, just pre-hash all output files from the previous run.  This
requires only slightly more hashing and avoids the extra run.