Better input path normalization
Previously, we named tasks by the os.realpath of the input file, but
TeX will name its output files based on the given path name, not any
resolution of the path name.  Instead, take the os.realpath of the
directory to normalize the *location* of the input file, but keep the
file component of the path name unchanged.