Record relative input/output paths from LaTeX when possible
This makes it possible to move a latexrun database (along with the
source files).  Technically, latex depends on the working directory,
but we don't record that (maybe we should).  This also makes clean
much safer after a database move, since it won't try to delete files
in the original directory.