Take object directory argument; change default to latex.out
Previously, this was named latexrun, but this causes problems if
people try to keep a local copy of the script or have . in their PATH.
Change the name to latex.out and provide a command-line argument to
change it.  Put the dependency database back in .latexrun.db so its
location doesn't depend on the argument.