Make task inputs extensible
This replaces the previous dependency mechanism that treated files and
environment variables specially and required the task to stuff any
other dependencies in a clunky dictionary with a general-purpose,
extensible dependency mechanism based on "input functions".  Now each
input is recorded as a method name, arguments, and its output.  A task
is stable if these method calls all return the same value as the
previous run.

This makes pre-hashing clobbered inputs slightly more awkward, since
files are no longer special.  To handle this, this adds a
HashCache.clobber method to record files whose hashes are no longer
available.  This feels clunky, but works.