Fix \include of files in subdirectories
Since we direct TeX to write output files to a separate output
directory, any files \included from a subdirectory will cause TeX to
attempt to create an output file in a parallel subdirectory of the
output directory, which is likely not to exist.

Technically this is a TeX problem (anyone using -output-directory with
\include of files from subdirectories would encounter this), but since
latexrun opts the user in to -output-directory, it seems best for
latexrun to mask this problem by creating this parallel output tree.