Emit a dependency file that makefiles can include.
At the end of a build we output a dependency file in the object
directory containing a dependency list in make's format. We make no
effort to sort out the details of the dependencies: we just collect
all the file deps from the latex and bibtex tasks and make them
dependencies of the output pdf; the important thing is just that any
changes get detected so latex run is invoked, which will then figure
out the details.

If the output file is paper.pdf, we default to dumping the deps file
as latex.out/paper.pdf.d.