Turn absolute paths in warnings/errors into relative ones.
For Bazel LaTeX, we store a copy of the texmf into the build
environment. The path of the build environment may be relatively long
and contain some randomness, depending on the build environment. Turn
such paths into relative ones, based on the current working directory to
make output more readable.