Fix '--color always' option when not stdout output
Bug:
If not stdout, program will auto-detect terminfo capabilities despite
the choice regarding the `--color` command line argument.  Because of
this, terminfo terminal is never initialized and the program breaks when
trying to output any color code.

This change forces terminfo proper initialization if `--color always` is
requested, even if `os.isatty()` is `False`.